chore: Move OCI code from spin-loader to spin-oci
